fixed userWater

+++ controllers/userWater.js
@@ -1,4 +1,3 @@
-const { User } = require("../models/user");
 const { WaterIntake } = require("../models/waterIntakeSchema");
 
 const currentDate = Date.now();
@@ -17,13 +16,10 @@ const getWaterIntake = async (req, res) => {
 
 const addWaterIntake = async (req, res, next) => {
   try {
-    const { value } = req.body;
-    const { _id } = req.user;
-    const user = await User.findOne({ _id });
-    const { _id: owner } = user;
+    const { value } = req.body; 
+    const { _id: owner } = req.user;    
     
-    const water = await WaterIntake.findOne({ owner, date: todayDate });
-    // console.log(water);
+    const water = await WaterIntake.findOne({ owner, date: todayDate });  
 
     if (!water) {
       const water = await WaterIntake.create({ owner, value });
@@ -42,6 +38,18 @@ const addWaterIntake = async (req, res, next) => {
 const deleteByIdWater = async (req, res) => {
   try {
     const { _id } = req.body;
+    const ownerId = req.user._id;
+
+    const waterRecord = await WaterIntake.findOne({ _id });
+
+    if (!waterRecord) {
+      return res.status(404).json({ message: "No water intake record found" });
+    }
+
+    if (String(waterRecord.owner) !== String(ownerId)) {
+      return res.status(403).json({ message: "Unauthorized to delete this record" });
+    }
+  
 
     const result = await WaterIntake.findOneAndDelete({ _id });
